Bionic dlmalloc: Reset a macro used in dlmalloc_usable_size().
When FOOTERS tuning option is not used in dlmalloc, synchronization
is based around a global 'gm' malloc_state or "mstate" bookkeeping
data structure. When FOOTERS is defined, there is an additional
integrity check on the mstate controlling dlmalloc() and dlfree(). In
dlmalloc_usable_size(), 'fm' therefore is a local variable of type
'mstate' when FOOTERS is defined but 'fm' is defined to 'gm' if
FOOTERS is not defined. For safety, need to #undef the macro 'fm'
after its use here when FOOTERS is not defined.

Change-Id: I1084552b636b7fec22106773c8f632fba983878c